# Optimize Validate operator to validate chunks in parallel