--- a/src/lib/operators/validate.cpp
+++ b/src/lib/operators/validate.cpp
@@ -57,64 +57,53 @@ std::shared_ptr<const Table> Validate::_on_execute(std::shared_ptr<TransactionCo
   DebugAssert(transaction_context, "Validate requires a valid TransactionContext.");
   DebugAssert(transaction_context->phase() == TransactionPhase::Active, "Transaction is not active anymore.");
 
-  const auto chunk_count = input_table_left()->chunk_count();
-
+  const auto in_table = input_table_left();
+  const auto chunk_count = in_table->chunk_count();
   const auto our_tid = transaction_context->transaction_id();
   const auto snapshot_commit_id = transaction_context->snapshot_commit_id();
 
   std::vector<std::shared_ptr<JobTask>> jobs;
-  std::vector<std::shared_ptr<std::vector<std::shared_ptr<Chunk>>>> job_results;
-  auto job_chunk_id_start = ChunkID{0};
-  auto job_row_count = uint32_t{0};
-  for (auto chunk_id = job_chunk_id_start; chunk_id < chunk_count; ++chunk_id) {
-    const auto chunk = input_table_left()->get_chunk(chunk_id);
-    job_row_count += chunk->size();
+  std::vector<std::shared_ptr<Chunk>> output_chunks;
+  output_chunks.reserve(chunk_count);
+  std::mutex output_mutex;
 
-    if(job_row_count >= Chunk::DEFAULT_SIZE || chunk_id == (chunk_count - 1)) {
-      bool execute_directly = job_chunk_id_start == 0 && chunk_id == (chunk_count - 1);
+  auto chunk_id_job_start = ChunkID{0};
+  auto row_count_job = uint32_t{0};
+  for (auto chunk_id = chunk_id_job_start; chunk_id < chunk_count; ++chunk_id) {
+    const auto chunk = in_table->get_chunk(chunk_id);
+    row_count_job += chunk->size();
 
-      auto output_chunks = std::make_shared<std::vector<std::shared_ptr<Chunk>>>();
-      output_chunks->reserve(chunk_id - job_chunk_id_start);
+    if(row_count_job >= Chunk::DEFAULT_SIZE || chunk_id == (chunk_count - 1)) {
+      // In case of one job only, execute directly.
+      bool execute_directly = chunk_id_job_start == 0 && chunk_id == (chunk_count - 1);
 
       if(execute_directly) {
-        _process_chunks(job_chunk_id_start, chunk_id, our_tid, snapshot_commit_id, output_chunks);
+        _validate_chunks(in_table, chunk_id_job_start, chunk_id, our_tid, snapshot_commit_id, output_chunks, output_mutex);
       } else {
-        jobs.push_back(std::make_shared<JobTask>([=, this] {
-          _process_chunks(job_chunk_id_start, chunk_id, our_tid, snapshot_commit_id, output_chunks);
+        jobs.push_back(std::make_shared<JobTask>([=, this, &output_chunks, &output_mutex] {
+          _validate_chunks(in_table, chunk_id_job_start, chunk_id, our_tid, snapshot_commit_id, output_chunks, output_mutex);
         }));
         jobs.back()->schedule();
 
-        job_chunk_id_start = chunk_id + 1;
-        job_row_count = uint32_t{0};
+        // Prepare next job
+        chunk_id_job_start = chunk_id + 1;
+        row_count_job = 0;
       }
-
-      job_results.push_back(output_chunks);
     }
   }
-
-  auto output_chunks = std::vector<std::shared_ptr<Chunk>>{};
-  output_chunks.reserve(chunk_count);
 
   if(jobs.size() > 0) {
     // Merge job results
     CurrentScheduler::wait_for_tasks(jobs);
-    for (const auto& chunk_vector : job_results) {
-      for (const auto& chunk : *chunk_vector) {
-        output_chunks.emplace_back(chunk);
-      }
-    }
-  } else if(job_results.size() > 0) {
-    Assert(job_results.size() == 1, "Direct execution should not produce multiple result vectors.");
-    output_chunks = *job_results.front();
   }
 
-  return std::make_shared<Table>(input_table_left()->column_definitions(), TableType::References, std::move(output_chunks));
+  return std::make_shared<Table>(in_table->column_definitions(), TableType::References, std::move(output_chunks));
 }
 
-void Validate::_process_chunks(const ChunkID chunk_id_start, const ChunkID chunk_id_end, const TransactionID our_tid, const TransactionID snapshot_commit_id, std::shared_ptr<std::vector<std::shared_ptr<Chunk>>> output_chunks) {
+void Validate::_validate_chunks(const std::shared_ptr<const Table> in_table, const ChunkID chunk_id_start, const ChunkID chunk_id_end, const TransactionID our_tid, const TransactionID snapshot_commit_id, std::vector<std::shared_ptr<Chunk>> &output_chunks, std::mutex &output_mutex) {
 
   for (auto chunk_id = chunk_id_start; chunk_id <= chunk_id_end; ++chunk_id) {
-    const auto chunk_in = input_table_left()->get_chunk(chunk_id);
+    const auto chunk_in = in_table->get_chunk(chunk_id);
     Assert(chunk_in, "Did not expect deleted chunk here.");  // see #1686
 
     Segments output_segments;
@@ -171,7 +160,7 @@ void Validate::_process_chunks(const ChunkID chunk_id_start, const ChunkID chunk
 
       // Otherwise we have a Value- or DictionarySegment and simply iterate over all rows to build a poslist.
     } else {
-      referenced_table = input_table_left();
+      referenced_table = in_table;
       DebugAssert(chunk_in->has_mvcc_data(), "Trying to use Validate on a table that has no MVCC data");
       const auto mvcc_data = chunk_in->get_scoped_mvcc_data_lock();
       pos_list_out->guarantee_single_chunk();
@@ -192,7 +181,8 @@ void Validate::_process_chunks(const ChunkID chunk_id_start, const ChunkID chunk
     }
 
     if (!pos_list_out->empty() > 0) {
-      output_chunks->emplace_back(std::make_shared<Chunk>(output_segments));
+      std::lock_guard<std::mutex> lock(output_mutex);
+      output_chunks.emplace_back(std::make_shared<Chunk>(output_segments));
     }
   }
 }