--- a/src/lib/operators/validate.cpp
+++ b/src/lib/operators/validate.cpp
@@ -95,14 +95,14 @@ std::shared_ptr<const Table> Validate::_on_execute(std::shared_ptr<TransactionCo
     }
   }
 
-  if (jobs.size() > 0) {
+  if (!jobs.empty()) {
     CurrentScheduler::wait_for_tasks(jobs);
   }
 
   return std::make_shared<Table>(in_table->column_definitions(), TableType::References, std::move(output_chunks));
 }
 
-void Validate::_validate_chunks(const std::shared_ptr<const Table> in_table, const ChunkID chunk_id_start,
+void Validate::_validate_chunks(const std::shared_ptr<const Table> &in_table, const ChunkID chunk_id_start,
                                 const ChunkID chunk_id_end, const TransactionID our_tid,
                                 const TransactionID snapshot_commit_id,
                                 std::vector<std::shared_ptr<Chunk>>& output_chunks, std::mutex& output_mutex) {