--- a/src/lib/operators/validate.hpp
+++ b/src/lib/operators/validate.hpp
@@ -26,6 +26,8 @@ class Validate : public AbstractReadOnlyOperator {
   static bool is_row_visible(TransactionID our_tid, CommitID snapshot_commit_id, const TransactionID row_tid,
                              const CommitID begin_cid, const CommitID end_cid);
 
+  void _process_chunks(const ChunkID chunk_id_start, const ChunkID chunk_id_end, const TransactionID our_tid, const TransactionID snapshot_commit_id, std::vector<std::shared_ptr<Chunk>>& output_chunks);
+
  protected:
   std::shared_ptr<const Table> _on_execute(std::shared_ptr<TransactionContext> transaction_context) override;
   std::shared_ptr<const Table> _on_execute() override;