Currently, the `Validate` operator processes all chunks sequentially. Especially for large input tables this can be time-consuming.

This PR parallelizes the validation process on a chunk-level. Chunks of `100.000` rows and more are handled by individual`JobTask`s. 

Running the TPC-DS benchmark, we get the following results:

```
+----------------+-----------------+------+-----------------+------+------------+-----------------------------------------------+
| Benchmark      | prev. iter/s    | runs | new iter/s      | runs | change [%] |               p-value (significant if <0.001) |
+----------------+-----------------+------+-----------------+------+------------+-----------------------------------------------+
| 01             | 1.23315536976   | 74   | 1.33324241638   | 80   | +8%        |                                        0.0000 |
| 07             | 0.349959135056  | 21   | 0.533293247223  | 32   | +52%       |                   (run time too short) 0.0000 |
| 09             | 0.149977818131  | 9    | 0.233311876655  | 14   | +56%       | (run time too short) (not enough runs) 0.0000 |
| 10             | 0.0166652090847 | 1    | 0.0166653320193 | 1    | +0%        |    (run time too short) (not enough runs) nan |
| 13             | 0.149983122945  | 9    | 0.18331451714   | 11   | +22%       | (run time too short) (not enough runs) 0.0000 |
| 15             | 0.933237314224  | 56   | 1.59984433651   | 96   | +71%       |                                        0.0000 |
| 17             | 0.383291661739  | 23   | 0.633280575275  | 38   | +65%       |                   (run time too short) 0.0000 |
| 25             | 0.533257484436  | 32   | 1.11652958393   | 67   | +109%      |                   (run time too short) 0.0000 |
| 26             | 0.666588366032  | 40   | 0.96658462286   | 58   | +45%       |                   (run time too short) 0.0000 |
| 28             | 3.09959936142   | 186  | 3.19959402084   | 192  | +3%        |                   (run time too short) 0.0000 |
| 29             | 0.466603755951  | 28   | 0.849895179272  | 51   | +82%       |                   (run time too short) 0.0000 |
| 31             | 0.0666584670544 | 4    | 0.0833244845271 | 5    | +25%       | (run time too short) (not enough runs) 0.0176 |
| 34             | 0.399941444397  | 24   | 0.633297622204  | 38   | +58%       |                   (run time too short) 0.0000 |
| 39a            | 0.0333290807903 | 2    | 0.0333317145705 | 2    | +0%        | (run time too short) (not enough runs) 0.0054 |
| 39b            | 0.0333289951086 | 2    | 0.033331759274  | 2    | +0%        | (run time too short) (not enough runs) 0.0223 |
| 41             | 0.749932944775  | 45   | 0.74994623661   | 45   | +0%        |                                        0.9887 |
| 42             | 0.599925875664  | 36   | 1.39983057976   | 84   | +133%      |                                        0.0000 |
| 43             | 0.11665546149   | 7    | 0.133323639631  | 8    | +14%       | (run time too short) (not enough runs) 0.0000 |
| 45             | 1.5164488554    | 91   | 2.18314504623   | 131  | +44%       |                                        0.0000 |
| 48             | 0.166645333171  | 10   | 0.199988186359  | 12   | +20%       |                   (run time too short) 0.0000 |
| 50             | 0.566587567329  | 34   | 1.11655521393   | 67   | +97%       |                   (run time too short) 0.0000 |
| 62             | 0.399958074093  | 24   | 0.433306783438  | 26   | +8%        |                                        0.0000 |
| 65             | 0.349954247475  | 21   | 0.516615152359  | 31   | +48%       |                   (run time too short) 0.0000 |
| 69             | 0.566586494446  | 34   | 1.04996812344   | 63   | +85%       |                   (run time too short) 0.0000 |
| 73             | 0.533263146877  | 32   | 1.06656324863   | 64   | +100%      |                                        0.0000 |
| 79             | 0.199974954128  | 12   | 0.249988943338  | 15   | +25%       |                   (run time too short) 0.0000 |
| 81             | 2.13304376602   | 128  | 2.19980335236   | 132  | +3%        |                   (run time too short) 0.0000 |
| 83             | 2.38330936432   | 143  | 2.36641764641   | 142  | -1%        |                   (run time too short) 0.8066 |
| 88             | 0.349989384413  | 21   | 0.49996316433   | 30   | +43%       |                   (run time too short) 0.0000 |
| 93             | 0.0833239108324 | 5    | 0.0833271816373 | 5    | +0%        | (run time too short) (not enough runs) 0.0026 |
| 96             | 0.466610074043  | 28   | 0.849919676781  | 51   | +82%       |                   (run time too short) 0.0000 |
| 97             | 0.133318185806  | 8    | 0.149990394711  | 9    | +13%       | (run time too short) (not enough runs) 0.0000 |
| 99             | 0.199980735779  | 12   | 0.216657459736  | 13   | +8%        |                   (run time too short) 0.0000 |
| geometric mean |                 |      |                 |      | +35%       |                                               |
+----------------+-----------------+------+-----------------+------+------------+-----------------------------------------------+
```

**Options**:  `numactl -m0 -N0 ./hyriseBenchmarkTPCDS --scheduler --cache_binary_tables` 

In TPC-H the results are similar:

```
+----------------+-----------------+------+-----------------+------+------------+-----------------------------------------------+
| Benchmark      | prev. iter/s    | runs | new iter/s      | runs | change [%] |               p-value (significant if <0.001) |
+----------------+-----------------+------+-----------------+------+------------+-----------------------------------------------+
| TPC-H 01       | 0.0166658777744 | 1    | 0.016664981842  | 1    | -0%        |    (run time too short) (not enough runs) nan |
| TPC-H 02       | 0.349946945906  | 21   | 0.383284211159  | 23   | +10%       |                   (run time too short) 0.0000 |
| TPC-H 03       | 0.399943590164  | 24   | 0.649976074696  | 39   | +63%       |                   (run time too short) 0.0000 |
| TPC-H 04       | 0.333330184221  | 20   | 1.09989619255   | 66   | +230%      |                   (run time too short) 0.0000 |
| TPC-H 05       | 0.18330784142   | 11   | 0.333324968815  | 20   | +82%       |                   (run time too short) 0.0000 |
| TPC-H 06       | 4.24990701675   | 255  | 4.24992418289   | 255  | +0%        |                   (run time too short) 0.6056 |
| TPC-H 07       | 0.0999894589186 | 6    | 0.099983394146  | 6    | -0%        | (run time too short) (not enough runs) 0.0000 |
| TPC-H 08       | 0.266634732485  | 16   | 0.566645503044  | 34   | +113%      |                                        0.0000 |
| TPC-H 09       | 0.116653516889  | 7    | 0.149976223707  | 9    | +29%       | (run time too short) (not enough runs) 0.0000 |
| TPC-H 10       | 0.316632598639  | 19   | 0.383274137974  | 23   | +21%       |                   (run time too short) 0.0000 |
| TPC-H 11       | 0.84989386797   | 51   | 1.4164390564    | 85   | +67%       |                                        0.0000 |
| TPC-H 12       | 0.849866449833  | 51   | 1.13333070278   | 68   | +33%       |                   (run time too short) 0.0000 |
| TPC-H 13       | 0.316624581814  | 19   | 0.366608709097  | 22   | +16%       |                   (run time too short) 0.0000 |
| TPC-H 14       | 2.19999527931   | 132  | 2.24963259697   | 135  | +2%        |                                        0.0000 |
| TPC-H 15       | 1.68309319019   | 101  | 1.7833224535    | 107  | +6%        |                                        0.0000 |
| TPC-H 16       | 0.466603517532  | 28   | 0.516592085361  | 31   | +11%       |                   (run time too short) 0.0000 |
| TPC-H 17       | 0.0999837145209 | 6    | 0.133333101869  | 8    | +33%       | (run time too short) (not enough runs) 0.0000 |
| TPC-H 18       | 0.0666573271155 | 4    | 0.0666561573744 | 4    | -0%        | (run time too short) (not enough runs) 0.0000 |
| TPC-H 19       | 0.366655439138  | 22   | 0.466636151075  | 28   | +27%       |                   (run time too short) 0.0000 |
| TPC-H 20       | 0.283302009106  | 17   | 0.316615492105  | 19   | +12%       |                   (run time too short) 0.0000 |
| TPC-H 21       | 0.183303102851  | 11   | 0.283294171095  | 17   | +55%       |                   (run time too short) 0.0000 |
| TPC-H 22       | 1.43309664726   | 86   | 1.61663269997   | 97   | +13%       |                   (run time too short) 0.0000 |
| geometric mean |                 |      |                 |      | +31%       |                                               |
+----------------+-----------------+------+-----------------+------+------------+-----------------------------------------------+
```
**Options**: `numactl -m0 -N0 ./hyriseBenchmarkTPCH --scheduler --cache_binary_tables`