@Bensk1 2019-09-05
Just for completeness, these are the results we got when avoiding unnecessary validation for TPC-DS single-threaded: 

```
+----------------+--------------------+------+--------------------+------+------------+---------------------------------+
| Benchmark      | prev. iter/s       | runs | new iter/s         | runs | change [%] | p-value (significant if <0.001) |
+----------------+--------------------+------+--------------------+------+------------+---------------------------------+
| 01             | 14.3278636932373   | 860  | 15.5599336624146   | 934  | +9%        |                          0.0000 |
| 07             | 10.4895925521851   | 630  | 15.4648847579956   | 928  | +47%       |                          0.0000 |
| 09             | 4.14280605316162   | 249  | 5.85816240310669   | 352  | +41%       |                          0.0000 |
| 10             | 2.17812991142273   | 131  | 2.45920300483704   | 148  | +13%       |                          0.0000 |
| 13             | 4.5789098739624    | 275  | 5.81975698471069   | 350  | +27%       |                          0.0000 |
| 15             | 31.4809551239014   | 1889 | 50.8433074951172   | 3051 | +62%       |                          0.0000 |
| 17             | 9.30982112884521   | 559  | 17.8347587585449   | 1071 | +92%       |                          0.0000 |
| 25             | 13.0641613006592   | 784  | 28.1251373291016   | 1688 | +115%      |                          0.0000 |
| 26             | 23.2724514007568   | 1397 | 33.2905426025391   | 1998 | +43%       |                          0.0000 |
| 28             | 7.00755405426025   | 421  | 8.24041843414307   | 495  | +18%       |                          0.0000 |
| 29             | 13.4330682754517   | 806  | 26.8614139556885   | 1612 | +100%      |                          0.0000 |
| 31             | 0.389165043830872  | 24   | 0.412751495838165  | 25   | +6%        |                          0.0000 |
| 34             | 12.281304359436    | 738  | 23.5681819915771   | 1415 | +92%       |                          0.0000 |
| 35             | 0.0256988052278757 | 2    | 0.0261396951973438 | 2    | +2%        |        (not enough runs) 0.0052 |
| 41             | 47.1427345275879   | 2829 | 48.2791938781738   | 2897 | +2%        |                          0.0000 |
| 42             | 16.3785228729248   | 983  | 24.6462268829346   | 1479 | +50%       |                          0.0000 |
| 43             | 2.98187613487244   | 179  | 3.24914598464966   | 195  | +9%        |                          0.0000 |
| 45             | 43.2611923217773   | 2596 | 63.4126014709473   | 3805 | +47%       |                          0.0000 |
| 48             | 4.62313079833984   | 278  | 5.62748765945435   | 338  | +22%       |                          0.0000 |
| 50             | 18.5129852294922   | 1111 | 36.1979026794434   | 2172 | +96%       |                          0.0000 |
| 62             | 8.97665405273438   | 539  | 9.80152606964111   | 589  | +9%        |                          0.0000 |
| 65             | 4.24925374984741   | 255  | 5.35327291488647   | 322  | +26%       |                          0.0000 |
| 69             | 7.36688041687012   | 443  | 12.4734001159668   | 749  | +69%       |                          0.0000 |
| 73             | 19.3687210083008   | 1163 | 40.0111427307129   | 2401 | +107%      |                          0.0000 |
| 79             | 5.42030763626099   | 326  | 7.28280162811279   | 438  | +34%       |                          0.0000 |
| 81             | 22.9421787261963   | 1377 | 25.185754776001    | 1512 | +10%       |                          0.0000 |
| 83             | 23.6449737548828   | 1419 | 30.2118034362793   | 1813 | +28%       |                          0.0000 |
| 88             | 2.60969948768616   | 157  | 4.94121694564819   | 297  | +89%       |                          0.0000 |
| 93             | 0.417474150657654  | 26   | 0.628774404525757  | 38   | +51%       |                          0.0000 |
| 96             | 19.3237495422363   | 1160 | 40.389289855957    | 2424 | +109%      |                          0.0000 |
| 97             | 2.00449204444885   | 121  | 2.32112956047058   | 140  | +16%       |                          0.0000 |
| 99             | 4.5704665184021    | 275  | 4.68554639816284   | 282  | +3%        |                          0.0000 |
| geometric mean |                    |      |                    |      | +41%       |                                 |
+----------------+--------------------+------+--------------------+------+------------+---------------------------------+
```

I guess the parallel validation makes more sense for workloads where multiple chunks need to be validated because data is still modified.