@Bouncner 2019-09-05
From the description, I don’t understand why single-threaded is now faster. Can you elaborate on the changes a bit more?