@Bensk1 2019-09-05
Sorry for throwing it into the conversation without description, we talked about that during yesterday's master's project meeting. We will elaborate on that during the next fortnightly or in a separate PR.